[data-access] Move the data access types into osc-domain
DO NOT merge this commit in isolation - it will break the build. This commit contains compilation failures due to tight coupling between the domain types and the core of the server. These dependencies will be broken in a separate commit so that the changes are more easily visible. At that point the commits can either be squashed, or merged at the same time.