Remove m2e exclusions for bnd-maven-plugin (#311)
The exclusions for m2e were added when the Uber build was unmanageably
large, and slowed down the incremental build. The Uber bundle is now
much smaller and so the exclusion is no longer needed. Disabling the
exclusion is required in order to enable incremental development in
Bndtools 3.4 when it is released.

Signed-off-by: Neil Bartlett <neilx.bartlett@intel.com>